protobuf gen triggers less often
Currently it triggers every rust build, which slows things down a lot.

Instead, only trigger if generated code actually changes.

+++ src/rust/engine/process_execution/bazel_protos/build.rs
@@ -14,22 +14,16 @@ fn main() {
   mark_dir_as_rerun_trigger(&thirdpartyprotobuf);
 
   let grpcio_output_dir = PathBuf::from("src/gen");
-  make_clean_dir(&grpcio_output_dir);
-  generate_for_grpcio(&thirdpartyprotobuf, &grpcio_output_dir);
+  replace_if_changed(&grpcio_output_dir, |path| {
+    generate_for_grpcio(&thirdpartyprotobuf, path);
+    format(path);
+  });
 
   let tower_output_dir = PathBuf::from("src/gen_for_tower");
-  make_clean_dir(&tower_output_dir);
-  generate_for_tower(&thirdpartyprotobuf, tower_output_dir.clone());
-
-  let success = std::process::Command::new(env!("CARGO"))
-    .arg("fmt")
-    .arg("--package=bazel_protos")
-    .status()
-    .unwrap()
-    .success();
-  if !success {
-    panic!("Cargo formatting failed for generated protos. Output should be above.");
-  }
+  replace_if_changed(&tower_output_dir, |path| {
+    generate_for_tower(&thirdpartyprotobuf, path);
+    format(path);
+  });
 
   // Re-gen if, say, someone does a git clean on the gen dir but not the target dir. This ensures
   // generated sources are available for reading by programmers and tools like rustfmt alike.
@@ -171,7 +165,7 @@ fn generate_mod_rs(dir: &Path) -> Result<(), String> {
     .map_err(|err| format!("Failed to write mod.rs: {}", err))
 }
 
-fn generate_for_tower(thirdpartyprotobuf: &Path, out_dir: PathBuf) {
+fn generate_for_tower(thirdpartyprotobuf: &Path, out_dir: &Path) {
   tower_grpc_build::Config::new()
     .enable_server(true)
     .enable_client(true)
@@ -188,7 +182,7 @@ fn generate_for_tower(thirdpartyprotobuf: &Path, out_dir: PathBuf) {
     .unwrap_or_else(|e| panic!("protobuf compilation failed: {}", e));
 
   let mut dirs_needing_mod_rs = HashSet::new();
-  dirs_needing_mod_rs.insert(out_dir.clone());
+  dirs_needing_mod_rs.insert(out_dir.to_owned());
 
   for f in walkdir::WalkDir::new(std::env::var("OUT_DIR").unwrap())
     .into_iter()
@@ -206,7 +200,7 @@ fn generate_for_tower(thirdpartyprotobuf: &Path, out_dir: PathBuf) {
     // pop .rs
     parts.pop();
 
-    let mut dst = out_dir.clone();
+    let mut dst = out_dir.to_owned();
     for part in parts {
       dst.push(part);
       dirs_needing_mod_rs.insert(dst.clone());
@@ -220,16 +214,51 @@ fn generate_for_tower(thirdpartyprotobuf: &Path, out_dir: PathBuf) {
     std::fs::copy(f.path(), dst).unwrap();
   }
 
-  disable_clippy_in_generated_code(&out_dir).expect("Failed to strip clippy from generated code");
+  disable_clippy_in_generated_code(out_dir).expect("Failed to strip clippy from generated code");
 
   for dir in &dirs_needing_mod_rs {
     generate_mod_rs(dir).expect("Failed to write mod.rs");
   }
 }
 
-fn make_clean_dir(path: &Path) {
+///
+/// Replaces contents of directory `path` with contents of directory populated by passed function.
+/// Does not modify `path` if the contents are identical.
+///
+fn replace_if_changed<F: FnOnce(&Path)>(path: &Path, f: F) {
+  let tempdir = tempfile::TempDir::new().unwrap();
+  f(tempdir.path());
+  if !dir_diff::is_different(path, tempdir.path()).unwrap() {
+    return;
+  }
   if path.exists() {
     std::fs::remove_dir_all(path).unwrap();
   }
-  std::fs::create_dir_all(path).unwrap();
+  std::fs::create_dir_all(path.parent().unwrap()).unwrap();
+  std::fs::rename(tempdir.path(), path).unwrap()
+}
+
+fn format(path: &Path) {
+  let mut rustfmt = PathBuf::from(env!("CARGO"));
+  rustfmt.pop();
+  rustfmt.push("rustfmt");
+  let mut rustfmt_config = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+  rustfmt_config.pop(); // bazel_protos
+  rustfmt_config.pop(); // process_execution
+  rustfmt_config.push("rustfmt.toml");
+  for file in walkdir::WalkDir::new(path) {
+    let file = file.unwrap();
+    if file.file_type().is_file() {
+      let success = std::process::Command::new(&rustfmt)
+        .arg(file.path())
+        .arg("--config-path")
+        .arg(&rustfmt_config)
+        .status()
+        .unwrap()
+        .success();
+      if !success {
+        panic!("Cargo formatting failed for generated protos. Output should be above.");
+      }
+    }
+  }
 }
